alloy_transport
Module layers

Module for housing transport layers.

Structs

RateLimitRetryPolicy

RetryBackoffLayer RateLimitRetryPolicy implements RetryPolicy to determine whether to retry depending on the err.

RetryBackoffLayer

RetryBackoffLayer A Transport Layer that is responsible for retrying requests based on the error type. See TransportError.

RetryBackoffService

RetryBackoffLayer A Tower Service used by the RetryBackoffLayer that is responsible for retrying requests based on the error type. See TransportError and RateLimitRetryPolicy.

Traits

RetryPolicy

RetryBackoffLayer RetryPolicy defines logic for which TransportError instances should the client retry the request and try to recover from.