alloy_transport

Trait DualTransportHandler

pub trait DualTransportHandler<L, R> {
    // Required method
    fn call(
        &self,
        request: RequestPacket,
        left: L,
        right: R,
    ) -> TransportFut<'static>;
}

Trait that determines how to dispatch a request given two transports.

Required Methods

fn call( &self, request: RequestPacket, left: L, right: R, ) -> TransportFut<'static>

The type of the future returned by the transport.