Type Alias alloy_transport::TransportError

pub type TransportError<ErrResp = Box<RawValue>> = RpcError<TransportErrorKind, ErrResp>;

A transport error is an RpcError containing a [TransportErrorKind].

Aliased Type

enum TransportError<ErrResp = Box<RawValue>> {
    ErrorResp(ErrorPayload<ErrResp>),
    SerError(Error),
    DeserError {
        err: Error,
        text: String,
    },
    Transport(TransportErrorKind),
}

Variants

ErrorResp(ErrorPayload<ErrResp>)

Server returned an error response.

SerError(Error)

Tuple Fields

0: Error

The underlying serde_json error.

JSON serialization error.

DeserError

Fields

err: Error

The underlying serde_json error.

text: String

For deser errors, the text that failed to deserialize.

JSON deserialization error.

Transport(TransportErrorKind)

Tuple Fields

0: TransportErrorKind

The underlying transport error.

Transport error.

This variant is used when the error occurs during communication.