alloy_transport

Module layers

Module for housing transport layers.

Structs

FallbackLayer

Fallback layer for transparent transport failover. This layer will consume a list of transports to provide better availability and reliability.

FallbackService

The FallbackService consumes multiple transports and is able to query them in parallel, returning the first successful response.

RateLimitRetryPolicy

RateLimitRetryPolicy implements RetryPolicy to determine whether to retry depending on the err.

RetryBackoffLayer

A Transport Layer that is responsible for retrying requests based on the error type. See TransportError.

RetryBackoffService

A Tower Service used by the RetryBackoffLayer that is responsible for retrying requests based on the error type. See TransportError and RateLimitRetryPolicy.

Traits

RetryPolicy

RetryPolicy defines logic for which TransportError instances should the client retry the request and try to recover from.